refactor packaging in data layer
