+++ app/src/main/java/com/shevaalex/android/plugev/data/openchargemap/network/service/ChargingStationRetrofitService.kt
@@ -1,7 +1,7 @@
-package com.shevaalex.android.plugev.data.network.openchargemap.service
+package com.shevaalex.android.plugev.data.openchargemap.network.service
 
 import com.shevaalex.android.plugev.BuildConfig
-import com.shevaalex.android.plugev.data.network.openchargemap.model.ChargingStationNetworkDto
+import com.shevaalex.android.plugev.data.openchargemap.network.model.ChargingStationNetworkDto
 import com.shevaalex.android.plugev.domain.openchargemap.API_RESULT_LIMIT
 import retrofit2.http.GET
 import retrofit2.http.Query
