+++ app/src/test/java/com/shevaalex/android/plugev/data/openchargemap/ChargingStationRepositoryImplTest.kt
@@ -1,9 +1,9 @@
-package com.shevaalex.android.plugev.data.network.openchargemap
+package com.shevaalex.android.plugev.data.openchargemap
 
 import com.google.common.truth.Truth.assertThat
 import com.shevaalex.android.plugev.data.DataFactory
-import com.shevaalex.android.plugev.data.network.openchargemap.model.toDomainModel
-import com.shevaalex.android.plugev.data.network.openchargemap.service.ChargingStationRetrofitService
+import com.shevaalex.android.plugev.data.openchargemap.network.model.toDomainModel
+import com.shevaalex.android.plugev.data.openchargemap.network.service.ChargingStationRetrofitService
 import com.shevaalex.android.plugev.domain.openchargemap.model.DataResult
 import io.mockk.MockKAnnotations
 import io.mockk.coEvery
